chore(*): removed helpers file, fixed create tests
