+++ src/commands/create/create.ts
@@ -8,16 +8,16 @@ import {
   createAngularApp,
   createMinimalApp,
   createTemplateApp
-} from '../utils/utils'
-import { getFolders, getConfiguration } from '../utils/config-utils'
+} from '../../utils/utils'
+import { getFolders, getConfiguration } from '../../utils/config-utils'
 import {
   createFolderStructure,
   createFolder,
   createFile,
   fileExists
-} from '../utils/file'
+} from '../../utils/file'
 import chalk from 'chalk'
-import { Folder } from '../types'
+import { Folder } from '../../types'
 
 export async function create(parentFolderName: string, appType: string) {
   const configPath = '../config.json'
