+++ src/commands/create/spec/files-angular-app.json
