+++ src/commands/create/spec/files-minimal-app.json
