+++ src/commands/create/spec/files-react-app.json
