+++ test/commands/db.spec.ts
@@ -2,6 +2,7 @@ import dotenv from 'dotenv'
 import path from 'path'
 import { createFileStructure, buildDBs } from '../../src/main'
 import { createFolder, deleteFolder } from '../../src/utils/file'
+import { verifyStep } from '../../src/utils/test'
 import { generateTimestamp } from '../../src/utils/utils'
 
 describe('sasjs db', () => {
